+++ src/main/java/podo/odeego/domain/member/service/MemberService.java
@@ -1,4 +1,4 @@
-package podo.odeego.domain.user.service;
+package podo.odeego.domain.member.service;
 
 import java.util.Map;
 
@@ -8,8 +8,8 @@ import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import podo.odeego.domain.user.entity.Member;
-import podo.odeego.domain.user.repository.MemberRepository;
+import podo.odeego.domain.member.entity.Member;
+import podo.odeego.domain.member.repository.MemberRepository;
 
 @Service
 @Transactional
